requirements.txt
# Core AI/ML libraries
google-generativeai>=0.3.0
google-adk>=0.1.0

# PDF processing
PyPDF2>=3.0.0

# Image processing
Pillow>=10.0.0

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0

# Logging and observability
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
